Use template for menu entries
