+++ pt_miniscreen/pages/templates/menu/__init__.py
@@ -1,0 +1,1 @@
+from .page import Page
